Attach instance profile to bootstrap concourse.

+++ scripts/bootstrap/01-bootstrap-terraform.sh
@@ -3,8 +3,10 @@
 set -eux
 
 # Create AWS keypair
-aws ec2 delete-key-pair --key-name bootstrap
-aws ec2 create-key-pair --key-name bootstrap > ${WORKSPACE_DIR}/aws-keypair.json
+if [ ! -f ${WORKSPACE_DIR}/aws-keypair.json ]; then
+  aws ec2 delete-key-pair --key-name bootstrap
+  aws ec2 create-key-pair --key-name bootstrap > ${WORKSPACE_DIR}/aws-keypair.json
+fi
 
 # Provision terraform infrastructure
 terraform init ./terraform/stacks/bootstrap
